댓글기능 꾸현

+++ src/main/java/com/ssafy/happyhouse/controller/CommentController.java
@@ -2,6 +2,8 @@ package com.ssafy.happyhouse.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.ssafy.happyhouse.model.dto.Comment;
+import com.ssafy.happyhouse.model.dto.Member;
 import com.ssafy.happyhouse.service.CommentService;
 
 @Controller
@@ -23,18 +26,18 @@ public class CommentController {
 
 	@RequestMapping("/list")
 	@ResponseBody
-	private List<Comment> searchList(Model model) throws Exception {
-		return service.searchComment();
+	private List<Comment> searchList(int bno, Model model) throws Exception {
+		return service.searchComment(bno);
 	}
 
 	@RequestMapping("/insert")
 	@ResponseBody
-	private int insertComment(@RequestParam int bno, @RequestParam String content) throws Exception {
+	private int insertComment(@RequestParam int bno, @RequestParam String content, HttpSession session) throws Exception {
 		Comment comment = new Comment();
 		comment.setBno(bno);
 		comment.setContent(content);
-		comment.setUserid("ssesion");
-		System.out.println(comment.toString());
+		Member member = (Member) session.getAttribute("memberInfo");
+		comment.setUserid(member.getId());
 		return service.insertComment(comment);
 	}
 
@@ -49,7 +52,18 @@ public class CommentController {
 
 	@RequestMapping("/delete/{cno}")
 	@ResponseBody
-	private int deleteComment(@PathVariable int cno) throws Exception {
-		return service.deleteComment(cno);
+	private int deleteComment(@PathVariable int cno, HttpSession session) throws Exception {
+		List<Comment> comment = service.inquireComment(cno);
+		System.out.println(comment);
+		String commentId = comment.get(0).userid;
+		System.out.println(commentId);
+		Member member = (Member) session.getAttribute("memberInfo");
+		String id = member.getId();
+		System.out.println(id);
+		if(id.equals(commentId)) {
+			System.out.println("삭제중 ..");
+			return service.deleteComment(cno);
+		}
+		return 0;
 	}
 }
